backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
beautifulsoup4==4.12.2
sqlmodel==0.0.8
aiofiles==23.1.0
python-multipart==0.0.6
pydantic==1.10.11
requests==2.31.0
python-dotenv==1.0.1
openai==0.27.8
